+++ Mandelbrot_Set/mandelbrot.cu
@@ -1,17 +1,17 @@
 #include "mandelbrot.cuh"
 
-__global__ void mandelbrotKernel(unsigned char* output, int width, int height, float x_min, float x_max, float y_min, float y_max) {
+__global__ void mandelbrotKernel(unsigned char* output, int width, int height, double x_min, double x_max, double y_min, double y_max) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     int idy = blockIdx.y * blockDim.y + threadIdx.y;
     if (idx >= width || idy >= height) return;
 
-    float real = x_min + (x_max - x_min) * idx / (float)width;
-    float imag = y_min + (y_max - y_min) * idy / (float)height;
-    float z_real = 0, z_imag = 0;
+    double real = x_min + (x_max - x_min) * idx / (double)width;
+    double imag = y_min + (y_max - y_min) * idy / (double)height;
+    double z_real = 0, z_imag = 0;
     int iter = 0;
 
-    while (iter < MAX_ITER && z_real * z_real + z_imag * z_imag <= 4.0f) {
-        float temp = z_real * z_real - z_imag * z_imag + real;
+    while (iter < MAX_ITER && z_real * z_real + z_imag * z_imag <= 4.0) {
+        double temp = z_real * z_real - z_imag * z_imag + real;
         z_imag = 2 * z_real * z_imag + imag;
         z_real = temp;
         iter++;
@@ -32,7 +32,7 @@ __global__ void mandelbrotKernel(unsigned char* output, int width, int height, f
     }
 }
 
-void computeMandelbrot(unsigned char* d_output, int width, int height, float x_min, float x_max, float y_min, float y_max) {
+void computeMandelbrot(unsigned char* d_output, int width, int height, double x_min, double x_max, double y_min, double y_max) {
     dim3 blockSize(16, 16); // Back to original, reliable size
     dim3 gridSize((width + blockSize.x - 1) / blockSize.x, (height + blockSize.y - 1) / blockSize.y);
     mandelbrotKernel << <gridSize, blockSize >> > (d_output, width, height, x_min, x_max, y_min, y_max);
