+++ Mandelbrot_Set/utils.cpp
@@ -17,7 +17,7 @@ void updateTexture(sf::Texture& texture, unsigned char* h_output) {
     texture.update(h_output, WIDTH, HEIGHT, 0, 0);
 }
 
-void handleEvents(sf::RenderWindow& window, float& x_min, float& x_max, float& y_min, float& y_max, bool& needsUpdate, sf::Vector2i& lastMousePos, bool& isDragging) {
+void handleEvents(sf::RenderWindow& window, double& x_min, double& x_max, double& y_min, double& y_max, bool& needsUpdate, sf::Vector2i& lastMousePos, bool& isDragging) {
     sf::Event event;
 
     while (window.pollEvent(event)) {
@@ -27,17 +27,17 @@ void handleEvents(sf::RenderWindow& window, float& x_min, float& x_max, float& y
 
         // Zoom with mouse wheel
         if (event.type == sf::Event::MouseWheelScrolled) {
-            float zoom_factor = (event.mouseWheelScroll.delta > 0) ? 0.8f : 1.25f; // Zoom in: 0.8, out: 1.25
+            double zoom_factor = (event.mouseWheelScroll.delta > 0) ? 0.8f : 1.25f; // Zoom in: 0.8, out: 1.25
             sf::Vector2i mouse_pos = sf::Mouse::getPosition(window);
-            float mx = x_min + (x_max - x_min) * mouse_pos.x / WIDTH;
-            float my = y_min + (y_max - y_min) * mouse_pos.y / HEIGHT;
+            double mx = x_min + (x_max - x_min) * mouse_pos.x / WIDTH;
+            double my = y_min + (y_max - y_min) * mouse_pos.y / HEIGHT;
 
-            float new_width = (x_max - x_min) * zoom_factor;
-            float new_height = (y_max - y_min) * zoom_factor;
-            float new_x_min = mx - new_width * ((float)mouse_pos.x / WIDTH);
-            float new_x_max = mx + new_width * (1.0f - (float)mouse_pos.x / WIDTH);
-            float new_y_min = my - new_height * ((float)mouse_pos.y / HEIGHT);
-            float new_y_max = my + new_height * (1.0f - (float)mouse_pos.y / HEIGHT);
+            double new_width = (x_max - x_min) * zoom_factor;
+            double new_height = (y_max - y_min) * zoom_factor;
+            double new_x_min = mx - new_width * ((double)mouse_pos.x / WIDTH);
+            double new_x_max = mx + new_width * (1.0 - (double)mouse_pos.x / WIDTH);
+            double new_y_min = my - new_height * ((double)mouse_pos.y / HEIGHT);
+            double new_y_max = my + new_height * (1.0 - (double)mouse_pos.y / HEIGHT);
 
             x_min = new_x_min;
             x_max = new_x_max;
@@ -57,8 +57,8 @@ void handleEvents(sf::RenderWindow& window, float& x_min, float& x_max, float& y
         }
         if (event.type == sf::Event::MouseMoved && isDragging) {
             sf::Vector2i currentMousePos = sf::Mouse::getPosition(window);
-            float dx = (x_max - x_min) * (lastMousePos.x - currentMousePos.x) / WIDTH;
-            float dy = (y_max - y_min) * (lastMousePos.y - currentMousePos.y) / HEIGHT;
+            double dx = (x_max - x_min) * (lastMousePos.x - currentMousePos.x) / WIDTH;
+            double dy = (y_max - y_min) * (lastMousePos.y - currentMousePos.y) / HEIGHT;
 
             x_min += dx;
             x_max += dx;
